Move VAP fields into folder
