+++ src/applications/personalization/profile/components/personal-information/EmailInformationSection.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import React from 'react';
 
-import Email from './VAPEmail';
+import Email from './fields/VAPEmail';
 
 import { signInServiceName as signInServiceNameSelector } from 'platform/user/authentication/selectors';
